Module Bimage

Image processing library

v0.2.1 — homepage

exception Unsupported

Raised when attempting to use Char, Int8_signed, Int16_signed Bigarray types

type ('a, 'b) kind = ('a, 'b) Stdlib.Bigarray.kind

type u8 = Stdlib.Bigarray.int8_unsigned_elt

type u16 = Stdlib.Bigarray.int16_unsigned_elt

type i32 = Stdlib.Bigarray.int32_elt

type i64 = Stdlib.Bigarray.int64_elt

type f32 = Stdlib.Bigarray.float32_elt

type f64 = Stdlib.Bigarray.float64_elt

type c32 = Stdlib.Bigarray.complex32_elt

type c64 = Stdlib.Bigarray.complex64_elt

val u8 : (int, u8) kind

val u16 : (int, u16) kind

val i32 : (int32, i32) kind

val i64 : (int64, i64) kind

val f32 : (float, f32) kind

val f64 : (float, f64) kind

val c32 : (Stdlib.Complex.t, c32) kind

val c64 : (Stdlib.Complex.t, c64) kind

module Error : sig ... end

module Angle : sig ... end

The Angle type is used instead of a float whenever a function expects an angle argument to avoid ambiguity

module Point : sig ... end

Point is a 2 element float tuple used to perform calculations on (x, y) coordinates

module Color : sig ... end

Color contains methods for creating and inspecting color types

type gray = [

| `Gray

]

1-channels gray color type

type rgb = [

| `Rgb

]

3-channel RGB color type

type rgb_packed = [

| `Rgb_packed

]

type xyz = [

| `Xyz

]

3-channel XYZ color type

type yuv = [

| `Yuv

]

3-channel YUV color type

type rgba = [

| `Rgba

]

4-channel RGBA image

type any = [

| `Any

]

Any color image

val gray : gray Color.t

Gray color

val rgb : rgb Color.t

RGB color

val rgb_packed : rgb_packed Color.t

RGB packed into a signle channel

val xyz : xyz Color.t

XYZ color

val yuv : yuv Color.t

YUV color

val rgba : rgba Color.t

RGBA color

val color : int -> any Color.t

Generic color

module Kind : sig ... end

module Data : sig ... end

The Data module defines several operations on one dimensional image data

module Pixel : sig ... end

Pixels are float vectors used to store image data

module Image : sig ... end

The Image module defines a simple interface for manipulating image data

module Kernel : sig ... end

Kernels are used for filtering images using convolution

type ('a, 'b, 'c, 'd, 'e, 'f) filter = output:('d, 'e, 'f) Image.t -> ('a, 'b, 'c) Image.t array -> unit

module Transform : sig ... end

module Input : sig ... end

Defines the type used as input to operations

module Expr : sig ... end

Expr implements an operation combinator which can be used to build operations from low-level functions

module Op : sig ... end

Op is used to define pixel-level operations. These operations are performed on normalized floating-point values

module Hash : sig ... end

val (~@) : int -> Input.index