Module Bimage.Angle

type t

val of_degrees : float -> t

of_degrees deg creates new angle from deg degrees

val to_degrees : t -> float

to_degrees angle returns the value of the angle in degrees

val of_radians : float -> t

of_radians rad creates a new angle from rad radians

val to_radians : t -> float

to_radians angle returns the value of the angle in radians