Module Bimage.Kind

val max : ('a, 'b) kind -> 'a

max k returns the maximum normalized value for k

val min : ('a, 'b) kind -> 'a

min k returns the minimum normalized value for k

val max_f : ('a, 'b) kind -> float

max k returns the maximum normalized value for k as a float

val min_f : ('a, 'b) kind -> float

min k returns the minimum normalized value for k as a float

val to_float : ('a, 'b) kind -> 'a -> float

to_float k x converts a value of kind k to float

val of_float : ('a, 'b) kind -> float -> 'a

of_float k x converts a float to a value of kind k

val clamp : ('a, 'b) kind -> float -> float

Converts a float value to a value within the proper range for the given kind

val normalize : ('a, 'b) kind -> float -> float

Scales a value to the range 0.0-1.0

val denormalize : ('a, 'b) kind -> float -> float

Sclaes a value to the range (kind_min-kind_max)

val convert : from:('a, 'b) kind -> ('c, 'd) kind -> 'a -> 'c