Module Bimage

Image processing library

v0.5.0 — homepage

module type TYPE = Type.TYPE

module type COLOR = Color.COLOR

exception Unsupported

Raised when attempting to use Bigarray types other than u8, u16, f32, f64, i32, i64

module Error : sig ... end

module Angle : sig ... end

module Point : sig ... end

module Color : sig ... end

Color contains methods for creating and inspecting color types

type gray = Color.Gray.t

1-channels gray color type

type rgb = Color.Rgb.t

3-channel RGB color type

type xyz = [

| `Xyz

]

3-channel XYZ color type

type yuv = [

| `Yuv

]

3-channel YUV color type

type rgba = Color.Rgba.t

4-channel RGBA image

val gray : [ `Gray ] Color.t

Gray color

val rgb : [ `Rgb ] Color.t

RGB color

val xyz : xyz Color.t

XYZ color

val yuv : yuv Color.t

YUV color

val rgba : [ `Rgba ] Color.t

RGBA color

module Type : sig ... end

type u8 = Stdlib.Bigarray.int8_unsigned_elt

type u16 = Stdlib.Bigarray.int16_unsigned_elt

type i32 = Stdlib.Bigarray.int32_elt

type i64 = Stdlib.Bigarray.int64_elt

type f32 = Stdlib.Bigarray.float32_elt

type f64 = Stdlib.Bigarray.float64_elt

val u8 : (int, u8) Type.t

val u16 : (int, u16) Type.t

val i32 : (int32, i32) Type.t

val i64 : (int64, i64) Type.t

val f32 : (float, f32) Type.t

val f64 : (float, f64) Type.t

module Data : sig ... end

Data provides some extra operations for Bigarray.Array1

module Pixel : sig ... end

Pixel defines operations on individual pixels, which are stored as floatarray

module Image : sig ... end

Image defines an image type and functions that manipulate images directly

type 'c image_u8 = (int, u8, 'c) Image.t

type 'c image_u16 = (int, u16, 'c) Image.t

type 'c image_i32 = (int32, i32, 'c) Image.t

type 'c image_i64 = (int64, i64, 'c) Image.t

type 'c image_f32 = (float, f32, 'c) Image.t

type 'c image_f64 = (float, f64, 'c) Image.t

module Kernel : sig ... end

Convolution kernels

module Transform : sig ... end

module Input : sig ... end

Input contains methods for working with Expr/Filter inputs

module Expr : sig ... end

Expr define a combinator which can be used to build filters

module Filter : sig ... end

Filter contains functions that can be used to execute expressions

type ('a, 'b, 'c) filter = output:('a, 'b, 'c) Image.t -> Input.t -> unit

module Hash : sig ... end

Perceptual hashing